ReadlineManagerクラスの作成
